23/6 0009 - Excepciones de AlreadyExist y DoesNotExist reemplazadas por unas más genéricas + otros cambios

+++ src/main/java/com/utn/UDEE/service/InvoiceService.java
@@ -1,6 +1,8 @@
 package com.utn.UDEE.service;
 
-import com.utn.UDEE.exception.doesNotExist.UserDoesNotExist;
+import com.utn.UDEE.exception.DeleteException;
+import com.utn.UDEE.exception.ResourceAlreadyExistException;
+import com.utn.UDEE.exception.ResourceDoesNotExistException;
 import com.utn.UDEE.models.Address;
 import com.utn.UDEE.models.Invoice;
 import com.utn.UDEE.models.User;
@@ -8,9 +10,13 @@ import com.utn.UDEE.repository.InvoiceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.time.LocalDate;
+
+import static java.util.Objects.isNull;
 
 @Service
 public class InvoiceService {
@@ -30,10 +36,10 @@ public class InvoiceService {
         this.meterService = meterService;
     }
 
-    public Page<Invoice> getInvoiceBetweenDateByUser(Integer idUser, LocalDate since, LocalDate until, Pageable pageable) throws UserDoesNotExist {
+    public Page<Invoice> getInvoiceBetweenDateByUser(Integer idUser, LocalDate since, LocalDate until, Pageable pageable) throws ResourceDoesNotExistException {
         User user = userService.getUserById(idUser);
         if (user == null) {
-            throw new UserDoesNotExist("User doesn't exist");
+            throw new ResourceDoesNotExistException("User doesn't exist");
         } else {
             return invoiceRepository.findAllByUserAndDateBetween(user, since, until, pageable);
         }
@@ -43,13 +49,33 @@ public class InvoiceService {
         return invoiceRepository.findAllByUserAndPayed(user, false, pageable);
     }
 
-    public Invoice addNewInvoice(Invoice invoice) {
-        return invoiceRepository.save(invoice);
+    public Invoice addNewInvoice(Invoice invoice) throws ResourceAlreadyExistException {
+
+        Invoice invoiceExist = getInvoiceById(invoice.getId());
+        if(isNull(invoiceExist)){
+            return invoiceRepository.save(invoice);
+        }else{
+            throw new ResourceAlreadyExistException("Invoice already exists");
+        }
     }
 
-    public void deleteInvoiceById(Integer idInvoice) {
-        invoiceRepository.deleteById(idInvoice);
+
+    public void deleteInvoiceById(Integer idInvoice) throws ResourceDoesNotExistException, DeleteException {
+        Invoice invoice = getInvoiceById(idInvoice);
+        if(invoice == null){
+            throw new ResourceDoesNotExistException("Invoice doesn't exist");
+        }
+        if(isNull(invoice.getUser())) {
+            invoiceRepository.deleteById(idInvoice);
+        } else {
+            throw new DeleteException("It cannot be deleted because another object depends on it");
+        }
     }
+
+    public Invoice getInvoiceById(Integer idInvoice) {
+        return invoiceRepository.findById(idInvoice).orElseThrow(()->new HttpClientErrorException(HttpStatus.NOT_FOUND));
+    }
+
 
     public Page<Invoice> getAllUnpaidByAddress(Integer idAddress, Pageable pageable) {
         Address address = addressService.getAddressById(idAddress);
@@ -58,5 +84,11 @@ public class InvoiceService {
 
     public Page<Invoice> getAllInvoices(Pageable pageable) {
         return invoiceRepository.findAll(pageable);
+    }
+
+    public Page<Invoice> getAllInvoicesByUser(Integer idUser, Pageable pageable) {
+        User user = userService.getUserById(idUser);
+
+        return invoiceRepository.findAllByUser(user, pageable);
     }
 }
